Added all files
All files for the program are added. Demo5 is to be run in order to access StackUtils, which contains the core logic. curlytestreal.txt is the name of the file the user must type in to run the test. The user may also create their own .txt file with curly braces (either exclusively or otherwise) to run with the program and test if they are balanced.